backend/requirements.txt
asgiref==3.2.10
attrs==19.3.0
cffi==1.14.0
cryptography==2.9.2
Django==3.1
django-dirtyfields==1.4
djangorestframework==3.11.1
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.9.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wcwidth==0.2.5
